Backend login is nearly done
Backend login will prompt the SQL database for a list of users and finds
the user that matches the username.
If there is no user that matches the username, it will return a 404 not
found.
If there is but password doesn't match, it returns data along with
unauthenticated.
If there is and password matches, it returns data along with
authenticated.
Still need to add student/bank info to the backend response